requirements.txt
Flask==2.0.1
Flask-CORS==3.0.10
mysql-connector-python==8.0.26
gunicorn==20.1.0
python-dotenv==0.19.0
Werkzeug==2.0.1
cryptography==3.4.7